add Reset Function & draw CircumscribedCircle
리셋기능과 외접원 그리기 추가

+++ DrawCircle/PointCircleRadiusDlg.h
@@ -12,7 +12,7 @@ public:
 	PointCircleRadiusDlg(CWnd* pParent = nullptr);   // 표준 생성자입니다.
 	virtual ~PointCircleRadiusDlg();
 
-	float m_Radius = 0.0f;
+	int m_Radius = 0;
 
 // 대화 상자 데이터입니다.
 #ifdef AFX_DESIGN_TIME
